Reorganize and make into a platformio library
