--- example_programs/example_programs.h
@@ -1,6 +1,0 @@
-#ifndef EXAMPLE_PROGRAMS_H
-#define EXAMPLE_PROGRAMS_H
-
-void test_all_features();
-
-#endif /* EXAMPLE_PROGRAMS_H */
