+++ library.json
@@ -1,0 +1,34 @@
+{
+    "name": "Y-Badge",
+    "version": "1.0.0",
+    "description": "A library for interacting with version 2 and 3 of the Y-Badge hardware",
+    "keywords": "education, y-badge",
+    "repository": {
+        "type": "git",
+        "url": "https://github.com/y-badge/lib-y-badge.git"
+    },
+    "authors": [
+        {
+            "name": "Philip Lundrigan",
+            "email": "lundrigan@byu.edu",
+            "url": "https://philip.lundrigan.org",
+            "maintainer": true
+        },
+        {
+            "name": "Jeff Goeders",
+            "email": "jgoeders@byu.edu",
+            "url": "https://jeffgoeders.com"
+        }
+    ],
+    "license": "MIT",
+    "homepage": "https://y-badge.github.io",
+    "dependencies": {
+        "adafruit/Adafruit NeoPixel": "^1.12.2"
+    },
+    "frameworks": [
+        "arduino"
+    ],
+    "platforms": [
+        "espressif32"
+    ]
+}
